[kbn-pm] Default build and watch commands to use bazel (#105674)
