+++ packages/kbn-pm/src/commands/watch.ts
@@ -6,80 +6,23 @@
  * Side Public License, v 1.
  */
 
-import dedent from 'dedent';
-import { CliError } from '../utils/errors';
-import { log } from '../utils/log';
-import { parallelizeBatches } from '../utils/parallelize';
-import { ProjectMap, topologicallyBatchProjects } from '../utils/projects';
-import { waitUntilWatchIsReady } from '../utils/watch';
+import { runIBazel } from '../utils/bazel';
 import { ICommand } from './';
 
-/**
- * Name of the script in the package/project package.json file to run during `kbn watch`.
- */
-const watchScriptName = 'kbn:watch';
-
-/**
- * Name of the Kibana project.
- */
-const kibanaProjectName = 'kibana';
-
-/**
- * Command that traverses through list of available projects/packages that have `kbn:watch` script in their
- * package.json files, groups them into topology aware batches and then processes theses batches one by one
- * running `kbn:watch` scripts in parallel within the same batch.
- *
- * Command internally relies on the fact that most of the build systems that are triggered by `kbn:watch`
- * will emit special "marker" once build/watch process is ready that we can use as completion condition for
- * the `kbn:watch` script and eventually for the entire batch. Currently we support completion "markers" for
- * `webpack` and `tsc` only, for the rest we rely on predefined timeouts.
- */
 export const WatchCommand: ICommand = {
-  description:
-    'Runs `kbn:watch` script for every project (only works on packages not using Bazel yet)',
+  description: 'Runs a build in the Bazel built packages and keeps watching them for changes',
   name: 'watch',
 
-  async run(projects, projectGraph) {
-    log.warning(dedent`
-      We are migrating packages into the Bazel build system. If the package you are trying to watch
-      contains a BUILD.bazel file please just use 'yarn kbn watch-bazel'
-    `);
+  async run(projects, projectGraph, { options }) {
+    const runOffline = options?.offline === true;
 
-    const projectsToWatch: ProjectMap = new Map();
-    for (const project of projects.values()) {
-      // We can't watch project that doesn't have `kbn:watch` script.
-      if (project.hasScript(watchScriptName)) {
-        projectsToWatch.set(project.name, project);
-      }
-    }
-
-    if (projectsToWatch.size === 0) {
-      throw new CliError(
-        `There are no projects to watch found. Make sure that projects define 'kbn:watch' script in 'package.json'.`
-      );
-    }
-
-    const projectNames = Array.from(projectsToWatch.keys());
-    log.info(`Running ${watchScriptName} scripts for [${projectNames.join(', ')}].`);
-
-    // Kibana should always be run the last, so we don't rely on automatic
-    // topological batching and push it to the last one-entry batch manually.
-    const shouldWatchKibanaProject = projectsToWatch.delete(kibanaProjectName);
-
-    const batchedProjects = topologicallyBatchProjects(projectsToWatch, projectGraph);
-
-    if (shouldWatchKibanaProject) {
-      batchedProjects.push([projects.get(kibanaProjectName)!]);
-    }
-
-    await parallelizeBatches(batchedProjects, async (pkg) => {
-      const completionHint = await waitUntilWatchIsReady(
-        pkg.runScriptStreaming(watchScriptName, {
-          debug: false,
-        }).stdout! // TypeScript note: As long as the proc stdio[1] is 'pipe', then stdout will not be null
-      );
-
-      log.success(`[${pkg.name}] Initial build completed (${completionHint}).`);
-    });
+    // Call bazel with the target to build all available packages and run it through iBazel to watch it for changes
+    //
+    // Note: --run_output=false arg will disable the iBazel notifications about gazelle and buildozer when running it
+    // Can also be solved by adding a root `.bazel_fix_commands.json` but its not needed at the moment
+    await runIBazel(
+      ['--run_output=false', 'build', '//packages:build', '--show_result=1'],
+      runOffline
+    );
   },
 };
